requirements.txt
# ECHO v2 – Abhängigkeiten
# Python 3.11 oder 3.12
# Ollama muss laufen: ollama serve && ollama pull qwen2.5:3b

nicegui==2.14.0
chromadb==0.5.11
sentence-transformers==3.3.1
ollama==0.4.7
einops>=0.8.0
pydantic==2.10.6
python-dateutil==2.9.0

# PDF-Export (optional — benötigt libpango + libcairo)
weasyprint==62.3

# Tray + Overlay (optional — für tray_overlay.py)
pillow>=10.0.0
pystray>=0.19.5
pywebview>=5.4
pynput>=1.7.6

# Clipboard für echo_to_claude.py (optional)
pyperclip>=1.8.0
